Externalisation des listes chainées
