+++ main.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "store.h"
+#include "linkedList.h"
 
 
 int main(int argc, char *argv[]){
